Data objects for Spyglass
This change switches Spyglass from storing data in large multi-level
dictionaries to using several data objects. The intent of this change is
to make management of Spyglass data easier for developers to understand.

Implementation of the data objects will be handled in a separate change
to keep code review more manageable.

Change-Id: I101ad1ccbd95822965b8da8b6a644522eb2908e7