Add site_config.yaml back into examples
Spyglass uses additional configuration files to add in any information
not included by a data source. This configuration file should be usable
independent of the data source used. This change adds the
site_config.yaml example file back into the examples folder after it was
initially removed in the project separation of spyglass-plugin-xls.

Change-Id: I9be76c89076327b8ff928aa3da6c91cca0397cc2