Use Pipfile.lock only in tox gates
Pipenv validates that the Pipfile.lock that it uses for installing
packages is not out of date before following through on execution. If it
is not, it will throw an error. This change adds an env that tells
Pipenv to ignore the Pipfile and only use Pipfile.lock so gates do not
fail due to out of date packages.

Change-Id: I2fa541b31e14561c7cefc330e7ef2cc0b0e4abc6